requirements.txt
torch>=2.5.0
transformers>=4.48.0
accelerate>=0.30.0
peft>=0.12.0
numpy>=1.24.0
pillow>=10.0.0
sentencepiece>=0.1.99
safetensors>=0.4.0
tqdm>=4.65.0
scikit-learn>=1.3.0
requests>=2.31.0